TASK: Removed old new year event
* replaced it with konjunkturzyklus wechseln

Addresses: #28
