+++ app/src/CoreGameLogic/CoreGameLogicApp.php
@@ -9,7 +9,7 @@ use Domain\CoreGameLogic\DrivingPorts\ForCoreGameLogic;
 use Domain\CoreGameLogic\Dto\ValueObject\GameId;
 use Domain\CoreGameLogic\EventStore\GameEvents;
 use Domain\CoreGameLogic\Feature\Initialization\InitializationCommandHandler;
-use Domain\CoreGameLogic\Feature\Jahreswechsel\JahreswechselCommandHandler;
+use Domain\CoreGameLogic\Feature\KonjunkturzyklusWechseln\JahreswechselCommandHandler;
 use Domain\CoreGameLogic\Feature\Spielzug\SpielzugCommandHandler;
 use Neos\EventStore\Helper\InMemoryEventStore;
 use Neos\EventStore\Model\EventStream\ExpectedVersion;
