+++ app/src/CoreGameLogic/Feature/KonjunkturzyklusWechseln/Command/KonjunkturzyklusWechseln.php
@@ -2,14 +2,14 @@
 
 declare(strict_types=1);
 
-namespace Domain\CoreGameLogic\Feature\Initialization\Command;
+namespace Domain\CoreGameLogic\Feature\KonjunkturzyklusWechseln\Command;
 
 use Domain\CoreGameLogic\CommandHandler\CommandInterface;
 use Domain\CoreGameLogic\Dto\ValueObject\Konjunkturzyklus;
 
 final readonly class KonjunkturzyklusWechseln implements CommandInterface
 {
-    public function __construct(public int $year, public Konjunkturzyklus $konjunkturzyklus)
+    public function __construct(public Konjunkturzyklus $konjunkturzyklus)
     {
     }
 }
