+++ app/src/CoreGameLogic/Feature/KonjunkturzyklusWechseln/Event/KonjunkturzyklusWechselExecuted.php
@@ -2,10 +2,11 @@
 
 declare(strict_types=1);
 
-namespace Domain\CoreGameLogic\Feature\Initialization\Event;
+namespace Domain\CoreGameLogic\Feature\KonjunkturzyklusWechseln\Event;
 
 use Domain\CoreGameLogic\Dto\Enum\Kompetenzbereiche;
 use Domain\CoreGameLogic\Dto\Enum\KonjunkturzyklusType;
+use Domain\CoreGameLogic\Dto\ValueObject\CurrentYear;
 use Domain\CoreGameLogic\Dto\ValueObject\Kategorie;
 use Domain\CoreGameLogic\Dto\ValueObject\Konjunkturzyklus;
 use Domain\CoreGameLogic\EventStore\GameEventInterface;
@@ -13,7 +14,7 @@ use Domain\CoreGameLogic\EventStore\GameEventInterface;
 final readonly class KonjunkturzyklusWechselExecuted implements GameEventInterface
 {
     public function __construct(
-        public int $year,
+        public CurrentYear $year,
         public Konjunkturzyklus $konjunkturzyklus,
     ) {
     }
@@ -21,7 +22,7 @@ final readonly class KonjunkturzyklusWechselExecuted implements GameEventInterfa
     public static function fromArray(array $values): GameEventInterface
     {
         return new self(
-            $values['year'],
+            year: new CurrentYear($values['year']),
             konjunkturzyklus: new Konjunkturzyklus(
                 KonjunkturzyklusType::fromString($values['konjunkturzyklus']['type']),
                 $values['konjunkturzyklus']['description'],
